requirements.txt
absl-py==1.4.0
accelerate==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.2.0
annotated-types==0.6.0
anthropic==0.31.2
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
autoattack @ git+https://github.com/fra31/auto-attack@a39220048b3c9f2cca9a4d3a54604793c68eca7e
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
blake3==1.0.2
bleach==6.1.0
boltons==23.0.0
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.26.0
colorama==0.4.6
colorful==0.5.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
compressed-tensors==0.8.1
contextlib2==21.6.0
contourpy==1.0.7
crcmod==1.7
cryptography==42.0.5
cycler==0.11.0
dashscope==1.20.12
datasets==2.20.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1727240596241/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
decord==0.6.0
depyf==0.18.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
einops==0.8.0
eval_type_backport==0.2.2
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fastapi==0.115.7
ffmpy==0.3.2
filelock==3.17.0
fonttools==4.39.2
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gast==0.6.0
gdown==5.1.0
geotorch==0.3.0
gguf==0.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.16.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.29.2
idna==2.10
importlib_metadata==7.1.0
importlib_resources==6.3.2
iniconfig==2.0.0
interegular==0.3.3
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1725050136642/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.5
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kaggle==1.6.6
kiwisolver==1.4.4
lark==1.2.2
linkify-it-py==2.0.3
lit==15.0.7
llvmlite==0.43.0
lm-format-enforcer==0.10.9
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.1
mistralai==1.8.2
ml-collections==0.1.1
ml_dtypes==0.5.0
mlc-ai-cu122==0.17.1
modelscope==1.11.0
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.19.0
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.0
ninja==1.11.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.26.1
openai==1.60.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.9.15
oss2==2.18.6
outlines==0.1.11
outlines_core==0.1.26
packaging==23.0
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
partial-json-parser==0.2.1.1.post5
patsy==0.5.3
peft==0.12.0
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==11.2.1
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.8.2
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.20.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.13.6
pycosat==0.6.6
pycountry==24.6.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.13
pynvml==11.5.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1725448950123/work
ray==2.31.0
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
rich==14.0.0
robustbench @ git+https://github.com/RobustBench/robustbench.git@46a91f44524133b2cd8f721ec7e73ecb63f17fc8
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.3
safetensors==0.5.2
scikit-learn==1.5.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sse-starlette==2.2.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.45.2
statsmodels==0.13.5
sympy==1.13.1
tabulate==0.9.0
tencentcloud-sdk-python==3.0.1226
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.9
together==1.5.17
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.1
torchdiffeq==0.2.3
torchsde==0.2.5
torchvision==0.20.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1724960687225/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
trampoline==0.1.2
transformers==4.40.2
triton==3.1.0
typer==0.15.1
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
uc-micro-py==1.0.3
umap-learn==0.5.7
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.29.1
watchfiles==0.22.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webdataset==0.2.86
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.17.2
xformers==0.0.28.post3
xgrammar==0.1.10
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zstandard==0.22.0
